pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cernoid"
version = "1.0.0"
description = "Face Recognition Identity Management System"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.3",
    "torch>=1.9.0",
    "dlib>=19.22.0",
    "python-multipart>=0.0.5",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=0.7.0",
    "python-dotenv>=0.19.0",
    "gTTS>=2.2.3"
]

[project.scripts]
cernoid = "cernoid.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/cernoid"]

[tool.hatch.metadata]
allow-direct-references = true 